CORE-5001 Use testing/test-serialization in ledger-consensual-impl
